New Crypto Listings Alert/requirements.txt
requests==2.25.1
python-dateutil==2.8.1
pathlib2==2.3.5
Markdown==3.3.4